Create Entering-DAOs-as-a-Writer.md
Shooting semi blindly here. Also realise I will need a sidebar entry to line 54. or 55. 
Have used this format to attempt to create clickable drop downs. Related to me and also to a full 10 page tutorial 🎉 
<details>
<summary>A little about Gnomeski </summary>
<br />
content.. blah blah to appear as clickable drop down i hope.. 
</details>
<p></p>
This is also used to insert a clickable link at the end of the playbook to a full length, 10 page tutorial, for folks who would like more depth. Done this way hopefully will be minimalist impact but, my detail will not be lost.